+++ src/js/utils/showAlert.mjs
@@ -1,0 +1,4 @@
+// Function to show an alert with a given message
+export const showAlert = (message) => {
+  alert(message);
+};
